% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{check_files}
\alias{check_files}
\title{check_files
Run basic checks on files for compare_source_code and compare_text}
\usage{
check_files(f1, f2)
}
\arguments{
\item{f1}{Vector of paths to files containing code chunks from solution 1 (e.g. student submissions)}

\item{f2}{Vector of paths to files containing code chunks from solution 2 (e.g. example solutions)}
}
\value{
TRUE if it passes - throw an error otherwise
}
\description{
check_files
Run basic checks on files for compare_source_code and compare_text
}
