% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_py_env}
\alias{clear_py_env}
\title{This clears the Python environment \code{py}.}
\usage{
clear_py_env(module)
}
\arguments{
\item{module}{the \code{py} module accessed via reticulate}
}
\value{
Nothing
}
\description{
It will keep important starting objects so \code{py} such as \code{r}and builtins.
}
\examples{
reticulate::py_run_string("x = 3")
# this removes the `x`
clear_py_env(py)

}
\keyword{internal}
