% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_source_code}
\alias{compare_source_code}
\title{compare_source_code
Compare source code files}
\usage{
compare_source_code(f1, f2)
}
\arguments{
\item{f1}{Vector of paths to files containing code chunks from solution 1 (e.g. student submissions)}

\item{f2}{Vector of paths to files containing code chunks from solution 2 (e.g. example solutions)}
}
\value{
A vector of similarity measures (between 0 and 1) for each pair of files in the list.
}
\description{
compare_source_code
Compare source code files
}
