% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_tags.R
\name{disable_all_tags}
\alias{disable_all_tags}
\title{Disable all html tags}
\usage{
disable_all_tags(ele)
}
\arguments{
\item{ele}{html tag element}
}
\description{
Method to disable all html tags to not allow users to interact with the html.
}
\examples{
# add an href to all a tags
disable_all_tags(
  htmltools::tagList(
    htmltools::a(),
    htmltools::a()
  )
)
}
