% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_stuff.R
\name{do_stuff}
\alias{do_stuff}
\title{Returns the number squared,
  except 42, which stays as 42}
\usage{
do_stuff(x)
}
\arguments{
\item{x}{input, must be numeric}
}
\value{
stuffed output
}
\description{
Returns the number squared,
  except 42, which stays as 42
}

