% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{duplicate_py_env}
\alias{duplicate_py_env}
\title{Create a duplicate of a Python environment}
\usage{
duplicate_py_env(module)
}
\arguments{
\item{module}{the \code{py} module accessed via \code{reticulate}}
}
\value{
a Python \code{Dict} or dictionary that is not converted to an R data type
}
\description{
Create a duplicate of a Python environment
}
\examples{
reticulate::py_run_string("x = 3")
new_py_envir <- duplicate_py_env(py)
new_py_envir$items()

}
\keyword{internal}
