% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{external_evaluator}
\alias{external_evaluator}
\title{External execution evaluator}
\usage{
external_evaluator(
  endpoint = getOption("tutorial.external.host",
    Sys.getenv("TUTORIAL_external_evaluator_HOST", NA)),
  max_curl_conns = 50
)
}
\arguments{
\item{endpoint}{The HTTP(S) endpoint to POST the exercises to}

\item{max_curl_conns}{The maximum number of simultaneous HTTP requests to the
endpoint.}
}
\description{
\lifecycle{experimental}
}
