% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback}
\alias{feedback}
\title{Provide exercise feedback}
\usage{
feedback(message, correct, type = c("auto", "success", "info", "warning",
  "error", "custom"), location = c("append", "prepend", "replace"))
}
\arguments{
\item{message}{Feedback message. This can be either plain text or a set of HTML tags
created using the \pkg{htmltools} package.}

\item{correct}{Was the answer to the exercise correct.}

\item{type}{Type of feedback. The standard types are "success", "info", "error", and "warning".
Each of these types will result in the feedback text being enclosed in a colored
panel. A value of "auto" will result in "success" when \code{correct == TRUE} and
"error" otherwise. Pass "custom" to pass the feedback message through without any 
decoration.}

\item{location}{Location of feedback relative to output.}
}
\description{
Provide exercise feedback
}

