% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{filesystem_storage}
\alias{filesystem_storage}
\title{Filesystem-based storage for tutor state data}
\usage{
filesystem_storage(dir, compress = TRUE)
}
\arguments{
\item{dir}{Directory to store state data within}

\item{compress}{Should \code{.rds} files be compressed?}
}
\value{
Storage handler suitable for \code{options(tutorial.storage = ...)}
}
\description{
Tutorial state storage handler that uses the filesystem
as a backing store. The directory will contain tutorial
state data partitioned by user_id, tutorial_id, and
tutorial_version (in that order)
}
