% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_tags.R
\name{finalize_question}
\alias{finalize_question}
\title{Finalize a question}
\usage{
finalize_question(ele)
}
\arguments{
\item{ele}{html tag element}
}
\value{
An \pkg{htmltools} HTML object with appropriately appended classes
such that a tutorial question is marked as the final answer.
}
\description{
Mark a question as finalized by adding a \code{question-final} class to the HTML
output at the top level, in addition to disabling all tags with
\code{\link[=disable_all_tags]{disable_all_tags()}}.
}
\examples{
# finalize the question UI
finalize_question(
  htmltools::div(
    class = "custom-question",
    htmltools::div("answer 1"),
    htmltools::div("answer 2")
  )
)

}
