% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_all_state_objects}
\alias{get_all_state_objects}
\title{Get state objects}
\usage{
get_all_state_objects(session, exercise_output = TRUE)
}
\arguments{
\item{session}{tutorial session object}

\item{exercise_output}{a logical. Defaults to TRUE.}
}
\value{
objects in session.
}
\description{
Get state objects
}
