% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizList.R
\name{getAnswerFlagById}
\alias{getAnswerFlagById}
\title{用于获取正确答案的数据}
\usage{
getAnswerFlagById(data, FInterId)
}
\arguments{
\item{FInterId}{内码}

\item{data2}{列表的数据}
}
\value{
返回值，内部函数，不再导出
}
\description{
用于获取正确答案的数据
}
\examples{
getAnswerFlagById(data2,1);
}
