% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizList.R
\name{getAnswerTextByID}
\alias{getAnswerTextByID}
\title{用于获取选项列表的文本}
\usage{
getAnswerTextByID(data, FInterId)
}
\arguments{
\item{data}{用于获取选择列表的数据}

\item{FInterId}{用于内码的数据}
}
\value{
返回相应的数据，内部函数，不再导出
}
\description{
用于获取选项列表的文本
}
\examples{
getAnswerTextByID(data2,1);
}
