% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizList.R
\name{getQuestionTextByID}
\alias{getQuestionTextByID}
\title{根据FInterId获取相应的名称}
\usage{
getQuestionTextByID(data, FInterId)
}
\arguments{
\item{FInterId}{列表的问题}

\item{data1}{问题列表的数据}
}
\value{
返回文本,内部函数不用于修改
}
\description{
根据FInterId获取相应的名称
}
\examples{
getQuestionTextByID(data1,1);
}
