% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizFromExcel.R
\name{getQuizFromExcel}
\alias{getQuizFromExcel}
\title{直接从Excel中直接生产相应的列表}
\usage{
getQuizFromExcel(file = "./data-raw/quizTemplate.xlsx", lang = "en",
  caption = "测试列表", maxCount = 20)
}
\arguments{
\item{file}{包含目录在内的完整的相对路径}

\item{lang}{语言，默认为en，其次是cn}

\item{caption}{自定义标题}

\item{maxCount}{定义可以选取的最大的行数}
}
\value{
返回一个测试列表
}
\description{
直接从Excel中直接生产相应的列表
}
\examples{
getQuizFromExcel(file='./data-raw/quizTemplate.xlsx',lang='en');
}
