% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizList.R
\name{getQuizList}
\alias{getQuizList}
\title{用于获取测试列表的主接口数据}
\usage{
getQuizList(data, caption = "测试列表")
}
\arguments{
\item{data}{数据，一个列表，包含questions与answers两个页签}

\item{caption}{自定义标题}
}
\value{
返回值
}
\description{
用于获取测试列表的主接口数据
}
\examples{
getQuizList(data);
}
