% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizTemplate.R
\name{getQuizTemplate}
\alias{getQuizTemplate}
\title{获取测试的模板数据，提供中英文2种格式}
\usage{
getQuizTemplate(path = "./", lang = "en")
}
\arguments{
\item{path}{文件的路径，默认为当前项目目录}

\item{lang}{语言的内容，默认为英文，也可以为中文，指定为cn即可}
}
\value{
实际没有返回值,后续考虑完善
}
\description{
获取测试的模板数据，提供中英文2种格式
}
\examples{
#getQuizTemplate(lang='cn');
}
