% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizFromExcel.R
\name{getSampleData}
\alias{getSampleData}
\title{用于从试题随机获取相应的数据}
\usage{
getSampleData(data, maxCount = 20)
}
\arguments{
\item{data}{符合模板格式的数据}

\item{maxCount}{最大的抽样行数}
}
\value{
返回经过随时抽样的数据结构
}
\description{
用于从试题随机获取相应的数据
}
\examples{
getSampeData(data,maxCount=20)
}
