% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A toy learner R package}
\usage{
hello(x)
}
\arguments{
\item{x}{The input variable}
}
\value{
The output from \code{\link{print}}
}
\description{
A toy learner R package
}
\examples{
hello('r')
\dontrun{
hello('stranger')
}
}
