% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{initialize}
\alias{initialize}
\title{Initialize tutor within a document}
\usage{
initialize()
}
\description{
Initialize tutor within a document
}
\details{
Tutor is automatically initialiazed whenever you use an exercise
or a question so calling this function explicilty is typically not
required. You might need it if you were using e.g. only the video
embedding feature of tutor without exercises and questions.
}
\keyword{internal}

