% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_tutor}
\alias{initialize_tutor}
\title{Initialize Tutor R Markdown Extensions}
\usage{
initialize_tutor()
}
\description{
One time initialization of R Markdown extensions required by the 
\pkg{tutor} package. This function is typically called automatically 
as a result of using exercises or questions.
}
