% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr-package.R
\docType{package}
\name{learnr-package}
\alias{learnr}
\alias{learnr-package}
\title{learnr: Interactive Tutorials for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Create interactive tutorials using R Markdown. Use a combination of narrative, figures, videos, exercises, and quizzes to create self-paced tutorials for learning about R and R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/learnr/}
  \item \url{https://github.com/rstudio/learnr}
  \item Report bugs at \url{https://github.com/rstudio/learnr/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@rstudio.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Authors:
\itemize{
  \item Barret Schloerke \email{barret@rstudio.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item JJ Allaire \email{jj@rstudio.com} [conceptor]
}

Other contributors:
\itemize{
  \item Alexander Rossell Hayes \email{alex.rossellhayes@rstudio.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [contributor]
  \item Nischal Shrestha \email{nischal@rstudio.com} (\href{https://orcid.org/0000-0003-3321-1712}{ORCID}) [contributor]
  \item Angela Li \email{angelali921@gmail.com} (vignette) [contributor]
  \item RStudio [copyright holder, funder]
  \item  Ajax.org B.V. (Ace library) [contributor, copyright holder]
  \item Zeno Rocha (clipboard.js library) [contributor, copyright holder]
  \item Nick Payne (Bootbox library) [contributor, copyright holder]
  \item Jake Archibald (idb-keyval library) [contributor, copyright holder]
  \item i18next authors (i18next library) [contributor, copyright holder]
}

}
\keyword{internal}
