% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_tags.R
\name{mutate_tags}
\alias{mutate_tags}
\alias{mutate_tags.default}
\alias{mutate_tags.list}
\alias{mutate_tags.shiny.tag}
\title{Mutate shiny html tags}
\usage{
mutate_tags(ele, selector, fn, ...)

\method{mutate_tags}{default}(ele, selector, fn, ...)

\method{mutate_tags}{list}(ele, selector, fn, ...)

\method{mutate_tags}{shiny.tag}(ele, selector, fn, ...)
}
\arguments{
\item{ele}{shiny html tag element}

\item{selector}{css selector string}

\item{fn}{function to execute when a match is found}

\item{...}{possible future parameter extension}
}
\description{
S3 method to recursively look for elements according to a basic css string.
This method should not be used publically until adopted by \code{htmltools}.
}
\examples{
# add an href to all a tags
mutate_tags(
  htmltools::tagList(
    htmltools::a(),
    htmltools::a()
  ),
  "a",
  function(a_tag) {
    htmltools::tagAppendAttributes(a_tag, href="http://rstudio.com")
  }
)
}
