% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{new_remote_evaluator}
\alias{new_remote_evaluator}
\title{Remote execution evaluator}
\usage{
new_remote_evaluator(
  endpoint = getOption("tutorial.remote.host"),
  max_curl_conns = 50
)
}
\arguments{
\item{endpoint}{The HTTP(S) endpoint to POST the exercises to}

\item{max_curl_conns}{The maximum number of simultaneous HTTP requests to the
endpoint.}
}
\description{
Remote execution evaluator
}
