% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qmd.R
\name{parse_qmd}
\alias{parse_qmd}
\title{parse_qmd
Extract R code and text from qmd file and export to temporary .R and .txt files}
\usage{
parse_qmd(fin)
}
\arguments{
\item{fin}{Path to an input file}
}
\value{
Returns a table with paths to the temporary .R and .txt files, with one
 row per challenge problem.
}
\description{
parse_qmd
Extract R code and text from qmd file and export to temporary .R and .txt files
}
\details{
This assumes submitted answers to challenge problems are encapsulated
 with tags starting with "<!-- Challenge", and that R code chunks start and end
  with "```". All other lines in the submitted answer are assumed to be text.
}
