% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_clear_env}
\alias{py_clear_env}
\title{This clears the Python environment \code{py}.}
\usage{
py_clear_env()
}
\value{
Nothing
}
\description{
It will keep important initial objects such as \code{py} (main module),
\code{r} (reticulate interface to R), and the \code{builtins} module.
}
\examples{
\dontrun{
reticulate::py_run_string("x = 3")
# this removes the `x`
py_clear_env()
}
}
\keyword{internal}
