% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_copy_global_env}
\alias{py_copy_global_env}
\title{Create a duplicate of a Python environment}
\usage{
py_copy_global_env()
}
\value{
a Python \code{Dict} or dictionary
}
\description{
Create a duplicate of a Python environment
}
\examples{
\dontrun{
reticulate::py_run_string("x = 3")
new_py_envir <- py_copy_global_env()
new_py_envir$items()
}
}
\keyword{internal}
