% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{question}
\alias{answer}
\alias{question}
\title{Tutorial quiz question}
\usage{
question(text, ..., type = c("auto", "single", "multiple"),
  correct = "Correct!", incorrect = "Incorrect.")

answer(text, correct = FALSE)
}
\arguments{
\item{text}{Question or option text}

\item{...}{One or more answers}

\item{type}{Type of quiz question. Typically this can be automatically determined
based on the provided answers Pass \code{"single"} to indicate that even though
multiple correct answers are specified that inputs which include only one correct
answer are still correct. Pass \code{"multiple"} to force the use of checkboxes
(as opposed to radio buttons) even though only once correct answer was provided.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults 
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect.")}
}
\description{
Add an interative multiple choice quiz question to a tutorial.
}

