% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_anybox.R
\name{question_anybox}
\alias{question_anybox}
\title{Anybox question}
\usage{
question_anybox(
  text,
  ...,
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE,
  min_right = 1,
  max_wrong = 0
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{answers and extra parameters passed onto \code{\link{question}}.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{TRUE}.}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}

\item{min_right}{Minimum number of correct options which must be selected.}

\item{max_wrong}{Maximum number of incorrect options which may be selected.}
}
\description{
Creates an anybox group tutorial quiz question.  The student may select one
or more checkboxes before submitting their answer. An alternative to the
checkbox group tutorial quiz question, if there are multiple correct answers,
you may choose the minimum number of correct responses and the maximum number
of incorrect responses required to successfully complete the question.
}
\details{
Correct options should have a message which will display if the question is
passed with missed options.
}
\examples{
question_anybox(
  "Select at least two toppings that belong on a Margherita Pizza (and no more than one that doesn't):",
  answer("tomato", correct = TRUE, message = "Tomatoes too!"),
  answer("mozzarella", correct = TRUE, message = "Don't forget the cheese!"),
  answer("basil", correct = TRUE, message = "Basil gives it a distinctive flavor!"),
  answer("extra virgin olive oil", correct = TRUE, message = "You need olive oil too!"),
  answer("pepperoni", message = "Pepperoni is a great topping! ... just not on a Margherita Pizza"),
  answer("onions", message = "Onions!? No and yuck!"),
  answer("bacon", message = "Bacon doesn't belong here!"),
  answer("spinach", message = "Spinach? With Olive Oil? Only if you're Popeye!"),
  random_answer_order = TRUE,
  allow_retry = TRUE,
  try_again = "Be sure to select all four toppings!",
  min_right = 2,
  max_wrong = 1
)
}
\seealso{
\code{\link{question_checkbox}} \code{\link{question_radio}},
\code{\link{question_text}}
}
