% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_methods.R
\name{question_is_correct_value}
\alias{question_is_correct_value}
\title{Question is correct value}
\usage{
question_is_correct_value(is_correct, messages, ...)
}
\arguments{
\item{is_correct}{boolean that determines if a question answer is correct}

\item{messages}{a list of messages to be displayed.  The type of message will be determined by the \code{is_correct} value.}

\item{...}{possible future parameter expansion}
}
\description{
Helper method to return
}
\examples{
# Radio button question implementation of `question_is_correct`
question_is_correct.radio <- function(question, input_answer, ...) {
  if (is.null(input_answer)) {
    showNotification("Please select an answer before submitting", type = "error")
    req(input_answer)
  }
  for (ans in question$answers) {
    if (as.character(ans$option) == input_answer) {
      return(question_is_correct_value(
        ans$is_correct,
        ans$message
      ))
    }
  }
  question_is_correct_value(FALSE, NULL)
}
}
