% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{quiz}
\alias{answer}
\alias{question}
\alias{quiz}
\title{<Add Title>}
\usage{
question(text, ..., correct = "Correct!", incorrect = "Incorrect!")

answer(text, correct = FALSE)
}
\arguments{
\item{...}{One or more quiz questions}

\item{caption}{Text caption}
}
\description{
<Add Description>
}

