% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizBatch.R
\name{quizList}
\alias{quizList}
\title{Tutorial quiz questions}
\usage{
quizList(questionSet, caption = "测试列表")
}
\arguments{
\item{caption}{Optional quiz caption (defaults to "Quiz")}

\item{text}{Question or option text}

\item{type}{Type of quiz question. Typically this can be automatically determined
based on the provided answers Pass \code{"single"} to indicate that even though
multiple correct answers are specified that inputs which include only one correct
answer are still correct. Pass \code{"multiple"} to force the use of checkboxes
(as opposed to radio buttons) even though only once correct answer was provided.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults 
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect.")}

\item{message}{Additional message to display along with correct/incorrect feedback.}

\item{...}{One or more questions or answers}

\item{allow_retry}{Allow retry for incorrect answers.}

\item{random_answer_order}{Display answers in a random order.}
}
\description{
Add interative multiple choice quiz questions to a tutorial.
}
\examples{
\dontrun{
question("What number is the letter A in the alphabet?",
  answer("8"),
  answer("14"),
  answer("1", correct = TRUE),
  answer("23"),
  incorrect = "See [here](https://en.wikipedia.org/wiki/English_alphabet) and try again.",
  allow_retry = TRUE
)

question("Where are you right now? (select ALL that apply)",
  answer("Planet Earth", correct = TRUE),
  answer("Pluto"),
  answer("At a computing device", correct = TRUE),
  answer("In the Milky Way", correct = TRUE),
  incorrect = paste0("Incorrect. You're on Earth, ",
                     "in the Milky Way, at a computer.")
)
}

}
