% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\docType{data}
\name{random_praise}
\alias{random_praise}
\alias{random_encouragement}
\alias{random_praises}
\title{Random praise and encouragement}
\format{An object of class \code{character} of length 10.}
\usage{
random_praise()

random_encouragement

random_praises

random_encouragement
}
\value{
Character string with a random saying
}
\description{
Random praises and encouragements sayings to compliment your question and quiz experience.
}
\keyword{datasets}
