% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{read_request}
\alias{read_request}
\title{Get session information}
\usage{
read_request(session, name, default = NULL)
}
\arguments{
\item{session}{tutorial session}

\item{name}{character string of item}

\item{default}{what to return if session does not exist}
}
\value{
Value of name from session. Defaults to NULL.
}
\description{
Get session information
}
