% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Record a user action}
\usage{
record(label, action, data)
}
\arguments{
\item{label}{Unique label for action}

\item{action}{Name of action}

\item{data}{Custom data for action}
}
\description{
Record an action performed by a user in a tutorial (e.g. answering a question,
taking a hint, etc.).
}

