% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{remote_evaluator}
\alias{remote_evaluator}
\title{Remote execution evaluator}
\usage{
remote_evaluator(
  endpoint = getOption("tutorial.remote.host",
    Sys.getenv("TUTORIAL_REMOTE_EVALUATOR_HOST", NA)),
  max_curl_conns = 50
)
}
\arguments{
\item{endpoint}{The HTTP(S) endpoint to POST the exercises to}

\item{max_curl_conns}{The maximum number of simultaneous HTTP requests to the
endpoint.}
}
\description{
Remote execution evaluator
}
