% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{render_safe}
\alias{render_safe}
\alias{run_safe}
\title{Render or Run documents in a new, safe R environment}
\usage{
render_safe(input, ..., show = TRUE)

run_safe(file, ..., show = TRUE)
}
\arguments{
\item{input, file}{Input file (R script, Rmd, or plain markdown).}

\item{...}{extra arguements to be passed to \code{rmarkdown::\link[rmarkdown]{render}} or
\code{rmarkdown::\link[rmarkdown]{run}}}

\item{show}{Logical, whether to show the standard output on the screen while the child process
is running. Defaults to \code{TRUE}.}
}
\description{
When rendering (or running) a document with R markdown, it inherits the current R Global environment.  This will produce unexpected behaviors, such as poisoning the R Global environment with existing variables.  By rendering the document in a new, safe R environment, a \emph{vanilla}, rendered document is produced.
}
