% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_runner.R
\name{run_exercise}
\alias{run_exercise}
\title{Run an Exercise}
\usage{
run_exercise(exercise)
}
\arguments{
\item{exercise}{Exercise input from client}
}
\value{
HTML UI with results of exercise
}
\description{
Run an Exercise
}
\keyword{internal}

