% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_tutorial}
\alias{run_tutorial}
\title{Run a tutorial}
\usage{
run_tutorial(name = NULL, package = NULL, shiny_args = NULL)
}
\arguments{
\item{name}{Tutorial name (subdirectory within \code{tutorials/}
directory of installed package).}

\item{package}{Name of package}

\item{shiny_args}{Additional arguments to forward to
\code{\link[shiny:runApp]{shiny::runApp}}.}
}
\description{
Run a tutorial provided by an installed R package.
}
\examples{
# display all "learnr" tutorials
available_tutorials("learnr")

# run basic example within learnr
\dontrun{
run_tutorial("hello", "learnr")
}

}
\seealso{
\code{\link{safe}} and \code{\link{available_tutorials}}
}
