% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{safe}
\alias{safe}
\title{Execute R code in a safe R environment}
\usage{
safe(expr, ..., show = TRUE, env = safe_env())
}
\arguments{
\item{expr}{expression that contains all the necessary library calls to
execute.  Expressions within callr do not inherit the existing,
loaded libraries.}
}
\description{
When rendering (or running) a document with R markdown, it inherits the
current R Global environment.  This will produce unexpected behaviors,
such as poisoning the R Global environment with existing variables.  By
rendering the document in a new, safe R environment, a \emph{vanilla},
rendered document is produced.
}
\details{
Using \code{safe} should only be necessary when locally deployed.
}
\examples{
\dontrun{
# Direct usage
safe({learnr::run_tutorial("slidy", package = "learnr")})

# Programmatic usage
library(rlang)

expr <- quote(learnr::run_tutorial("slidy", package = "learnr"))
safe(!!expr)

tutorial <- "slidy"
safe({learnr::run_tutorial(!!tutorial, package = "learnr")})
}
}
