% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_time.R
\name{show_time}
\alias{show_time}
\title{Function to show time at user defined time zone}
\usage{
show_time(time_zone = "GMT")
}
\arguments{
\item{time_zone}{Time zone variable, default "GMT"}
}
\value{
Time at user defined time zone
}
\description{
Function to show time at user defined time zone
}
\examples{
show_time("US/Pacific")
}
