% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{submissions_from_state_objects}
\alias{submissions_from_state_objects}
\title{Get submissions}
\usage{
submissions_from_state_objects(state_objects)
}
\arguments{
\item{state_objects}{state objects from tutorial}
}
\value{
answered questions
}
\description{
Get submissions
}
