% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-dependencies.R
\name{teachdown_html_dependency}
\alias{teachdown_html_dependency}
\title{Tutor HTML dependency}
\usage{
teachdown_html_dependency()
}
\description{
Tutor HTML dependency
}
\details{
HTML dependency for core teachdown JS and CSS. This should be included as a 
dependency for custom tutorial formats that wish to ensure that that
teachdown.js and teachdown.css are loaded prior their own scripts and stylesheets.
}
