% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-dependencies.R
\name{tutor_html_dependency}
\alias{tutor_html_dependency}
\title{Tutor HTML dependency}
\usage{
tutor_html_dependency()
}
\description{
Tutor HTML dependency
}
\details{
HTML dependency for core tutor JS and CSS. This should be included as a 
dependency for custom tutorial formats that wish to ensure that that
tutor.js and tutor.css are loaded prior their own scripts and stylesheets.
}

