% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tutor_options}
\alias{tutor_options}
\title{Set tutor options}
\usage{
tutor_options(exercise.cap = "Code", exercise.eval = FALSE,
  exercise.timelimit = 30, exercise.lines = NULL, exercise.checker = NULL)
}
\arguments{
\item{exercise.cap}{Caption for exercise chunk (defaults to \code{"Code")}.}

\item{exercise.eval}{Whether to pre-evaluate the exercise so the reader can 
see some default output (defaults to \code{FALSE}).}

\item{exercise.timelimit}{Number of seconds to limit execution time to 
(defaults to \code{30}).}

\item{exercise.lines}{Lines of code for exercise editor (defaults to the 
number of lines in the code chunk).}

\item{exercise.checker}{Function used to check exercise answers.}
}
\description{
Set various tutor options that control the display and evaluation of 
exercises.
}

