% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-dependencies.R
\name{tutorial_html_dependency}
\alias{tutorial_html_dependency}
\title{Tutorial HTML dependency}
\usage{
tutorial_html_dependency()
}
\value{
\pkg{learnr}'s HTML dependencies
}
\description{
HTML dependency for core tutorial JS and CSS. This should be included as a
dependency for custom tutorial formats that wish to ensure that that
\code{tutorial.js} and \code{tutorial.css} are loaded prior their own scripts and
stylesheets.
}
