% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-dependencies.R
\name{tutorial_i18n_custom_language}
\alias{tutorial_i18n_custom_language}
\title{Customize the Language of a learnr Tutorial}
\usage{
tutorial_i18n_custom_language(language = NULL, button = NULL, text = NULL)
}
\arguments{
\item{language}{Default language for the custom translation. Translateable
fields without a customization will inherit from the translations for this
language.}

\item{button}{A list of keys with translations for button text. One of
"runcode", "hints", "startover", "continue", "submitanswer", "previoustopic", or "nexttopic")`.}

\item{text}{A list of keys with translations for text elements. One of
"startover", "areyousure", "youmustcomplete", or "inthissection")`.}
}
\value{
Returns a list formatted for use with \pkg{learnr} translation
functionality.
}
\description{
The language of various elements of a learnr tutorial can be customized via
the \code{language} option of the \code{\link[=tutorial]{tutorial()}} format. To customize the
language used in one specific tutorial, the \code{language} option is recommended.
Authors who rely on many customizations or who are providing complete
language translations can use \code{tutorial_i18n_custom_language()} to setup
customization settings that can be written into a \code{.json} file and used
across many tutorials. See \verb{vignette("multilang", package = "learnr')} for
a complete overview of language customization options.
}
\examples{
tutorial_i18n_custom_language(
  language = "en",
  button = list(runcode = "Run!"),
  text = list(startover = "Restart!")
)

tutorial_i18n_custom_language("es", button = list(runcode = "Ejecutar"))

}
\seealso{
\verb{vignette("multilang", package = "learnr')}
}
