% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tutorial_options}
\alias{tutorial_options}
\title{Set tutorial options}
\usage{
tutorial_options(
  exercise.cap = NULL,
  exercise.eval = FALSE,
  exercise.timelimit = 30,
  exercise.lines = NULL,
  exercise.checker = NULL,
  exercise.error.check.code = NULL,
  exercise.completion = TRUE,
  exercise.diagnostics = TRUE,
  exercise.startover = TRUE,
  exercise.alert_class = "red",
  exercise.feedback_show = TRUE,
  exercise.code_show = TRUE,
  exercise.execution_error_message = NULL,
  exercise.gradethis_success_color = NULL,
  exercise.gradethis_info_color = NULL,
  exercise.gradethis_warning_color = NULL,
  exercise.gradethis_danger_color = NULL,
  exercise.gradethis_feedback_show = TRUE,
  exercise.gradethis_code_show = TRUE
)
}
\arguments{
\item{exercise.cap}{Caption for exercise chunk (defaults to the engine's icon or the combination of the engine and \code{" code"}).}

\item{exercise.eval}{Whether to pre-evaluate the exercise so the reader can
see some default output (defaults to \code{FALSE}).}

\item{exercise.timelimit}{Number of seconds to limit execution time to
(defaults to \code{30}).}

\item{exercise.lines}{Lines of code for exercise editor (defaults to the
number of lines in the code chunk).}

\item{exercise.checker}{Function used to check exercise answers
(e.g., \code{gradethis::grade_learnr()}).}

\item{exercise.error.check.code}{A string containing R code to use for checking
code when an exercise evaluation error occurs (e.g., \code{"gradethis::grade_code()"}).}

\item{exercise.completion}{Use code completion in exercise editors.}

\item{exercise.diagnostics}{Show diagnostics in exercise editors.}

\item{exercise.startover}{Show "Start Over" button on exercise.}

\item{exercise.alert_class, exercise.gradethis_info_color, exercise.gradethis_success_color, exercise.gradethis_warning_color, exercise.gradethis_danger_color}{Slug for the CSS class for \code{{learnr}} and \code{{gradethis}} message.
It can be one of \code{red}, \code{orange}, \code{purple}, \code{blue}, \code{violet}, \code{yellow}, \code{pink}, \code{green},
or \code{grey}. You can also implement your own CSS rule, in that case you need to define a
class that starts with \verb{alert-} (for example \code{alert-rainbow}).}

\item{exercise.feedback_show}{Should the \code{{learnr}} feedback be shown?}

\item{exercise.code_show}{Should \code{{learnr}} output code be shown?}

\item{exercise.execution_error_message}{What message should \code{{learnr}} print on error?}

\item{exercise.gradethis_feedback_show}{Should the \code{{gradethis}} feedback be shown?}

\item{exercise.gradethis_code_show}{Should \code{{gradethis}} output code be shown?}
}
\description{
Set various tutorial options that control the display and evaluation of
exercises.
}
