% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_package_dependencies.R
\name{tutorial_package_dependencies}
\alias{tutorial_package_dependencies}
\title{List Tutorial Dependencies}
\usage{
tutorial_package_dependencies(name, package)
}
\arguments{
\item{name}{The tutorial name.}

\item{package}{The \R package providing the tutorial.}
}
\description{
List the \R packages required to run a particular tutorial.
}
