% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_data_dir.R
\name{use_data_dir}
\alias{use_data_dir}
\title{Make files in a directory accessible to tutorials}
\usage{
use_data_dir(dir = "data")
}
\arguments{
\item{dir}{A directory}
}
\value{
Invisibly returns the specified directory
}
\description{
By default, \code{learnr} tutorials can access files in a \verb{data/} directory within
the same directory as the tutorial's R Markdown file. \code{use_data_dir()} allows
you to expose files in a different directory (e.g. the \verb{data/} directory of a
different tutorial) to avoid duplicating files.
}
\examples{
\dontrun{
use_data_dir("../other_tutorial/data")
}
}
