% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{use_remote_files}
\alias{use_remote_files}
\title{Make remote files accessible in a tutorial}
\usage{
use_remote_files(...)
}
\arguments{
\item{...}{Remote files to include in the tutorial.
Each file will be accessible at \verb{data/<name of input>}.
Unnamed arguments have their name set by default with \code{\link[=basename]{basename()}}.}
}
\value{
Sets \link[knitr:opts_chunk]{\code{knitr} chunk options}
}
\description{
By default, all files in the \verb{data/} directory alongside a tutorial will be
accessible within the tutorial.
\code{use_remote_files()} makes files stored remotely (e.g. online) accessible
as well.
}
\section{Usage in learnr tutorials}{


To use remote files in a \code{learnr} tutorial, you can either include a call to
\code{use_remote_files()} in your global setup chunk:\preformatted{```\{r setup, include = FALSE\}`r ''`
library(learnr)
use_remote_files(

)
```
}

Alternatively, you can call \code{random_phrases_add()} in a separate, standard
R chunk (with \code{echo = FALSE}):\preformatted{```\{r setup-phrases, echo = FALSE\}`r ''`
use_remote_file(
  "https://covidtracking.com/api/v1/states/daily.csv",
  "taxi.csv" = "https://s3.amazonaws.com/nyc-tlc/trip+data/yellow_tripdata_2020-01.csv"
)
```
}
}

\examples{
use_remote_file(
  "https://covidtracking.com/api/v1/states/daily.csv",
  "https://s3.amazonaws.com/nyc-tlc/trip+data/yellow_tripdata_2020-01.csv"
)

use_remote_file(
  "covid.csv" = "https://covidtracking.com/api/v1/states/daily.csv",
  "taxi.csv"  = "https://s3.amazonaws.com/nyc-tlc/trip+data/yellow_tripdata_2020-01.csv"
)


}
