% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_tutorials.R
\name{available_tutorials}
\alias{available_tutorials}
\title{List available tutorials}
\usage{
available_tutorials(package = NULL)
}
\arguments{
\item{package}{Name of package}
}
\value{
\code{available_tutorials()} returns a \code{data.frame} containing "package",
"name", "title", "description", "package_dependencies", "private", and
"yaml_front_matter".
}
\description{
List the tutorials that are currently available via installed R packages.
Or list the specific tutorials that are contained within a given R package.
}
\examples{
available_tutorials(package = "learnr")

}
