% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_tutorials.R
\name{available_tutorials}
\alias{available_tutorials}
\title{Run a tutorial}
\usage{
available_tutorials(package = NULL)
}
\arguments{
\item{package}{Name of package}
}
\value{
\code{available_tutorials} will return a \code{data.frame} containing "package", "name", "title", "description", "package_dependencies", "private", and "yaml_front_matter".
}
\description{
Run a tutorial which is contained within an R package.
}
\details{
Note that when running a tutorial Rmd file with \code{run_tutorial}
the tutorial Rmd should have already been rendered as part of the
development of the package (i.e. the corresponding tutorial .html file for
the .Rmd file must exist).
}
