% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exercise.R
\name{check_exercise}
\alias{check_exercise}
\title{Check the code from an Exercise}
\usage{
check_exercise(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to the Markdown file containing the RMarkdown.}

\item{verbose}{Should the test output information on the console?}
}
\value{
TRUE or FALSE invisibly.
}
\description{
This function will take all the chunks with a label that matches \code{setup} or
\code{-solution}, put them in a separate script and try to run them all.
This will allow teachers to check that their setup and solution chunks
contain valid code.
}
\examples{
if (interactive()){
  check_exercise("sandbox/sandbox.Rmd")
}
}
