% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_register_handler}
\alias{event_register_handler}
\alias{event_register_handler_once}
\title{Register an event handler callback}
\usage{
event_register_handler(event, callback)

event_register_handler_once(event, callback)
}
\arguments{
\item{event}{The name of an event.}

\item{callback}{A function to be invoked when an event with a specified name
occurs. The callback must take parameters \code{session}, \code{event}, and \code{data}.}
}
\value{
A function which, if invoked, will remove the callback.
}
\description{
Register an event handler on a per-tutorial basis. Handlers for an event will
be fired in the order that they were registered. If
\code{event_register_handler_once()} is used, the callback will execute only one
time.
}
