% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback}
\alias{feedback}
\title{Provide exercise feedback}
\usage{
feedback(message, type = c("success", "info", "warning", "error", "custom"),
  location = c("append", "prepend", "replace"))
}
\arguments{
\item{message}{Feedback message. This can be either plain text or a set of HTML tags
created using the \pkg{htmltools} package.}

\item{type}{Type of feedback. The standard types are "success", "info", "error", and "warning".
Each of these types will result in the feedback text being enclosed in a colored
panel. Pass "custom" to pass the feedback message through without any decoration.}

\item{location}{Location of feedback relative to output.}
}
\description{
Provide exercise feedback
}

