% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise-cache.R
\name{get_tutorial_info}
\alias{get_tutorial_info}
\title{Get information about the current tutorial}
\usage{
get_tutorial_info(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer.} Default is \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}}.}
}
\value{
Returns an ordinary list with the following elements:
\itemize{
\item \code{tutorial_id}: The ID of the tutorial, auto-generated or from the
\code{tutorial$id} key in the tutorial's YAML front matter.
\item \code{tutorial_version}: The tutorial's version, auto-generated or from the
\code{tutorial$version} key in the tutorial's YAML front matter.
\item \code{user_id}: The current user.
\item \code{learnr_version}: The current version of the running learnr package.
\item \code{language}: The current language of the tutorial, either as chosen by the
user or as specified in the \code{language} item of the YAML front matter.
}
}
\description{
Returns information about the current tutorial. Ideally the function should
be evaluated in a Shiny context, i.e. in a chunk with option
\code{context = "server"}. Note that the values of this function may change after
the tutorial is completely initialized. If called in a non-reactive context,
\code{get_tutorial_info()} will return default values that will most likely
correspond to the current tutorial.
}
\seealso{
\code{\link[=get_tutorial_state]{get_tutorial_state()}}
}
