% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Initialize Tutor R Markdown Extensions}
\usage{
initialize()
}
\description{
One time initialization of R Markdonw extensions required by the 
\pkg{tutor} package. This function is typically called automatically 
as a result of using exercises or questions.
}

