% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_tutorial}
\alias{initialize_tutorial}
\title{Initialize tutorial R Markdown extensions}
\usage{
initialize_tutorial()
}
\description{
One time initialization of R Markdown extensions required by the
\pkg{learnr} package. This function is typically called automatically
as a result of using exercises or questions.
}
