% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{quiz}
\alias{question}
\alias{quiz}
\title{<Add Title>}
\usage{
quiz(..., caption = NULL)

question(caption)
}
\arguments{
\item{...}{One or more quiz questions}

\item{caption}{Text caption}
}
\description{
<Add Description>
}

