% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{random_praise}
\alias{random_praise}
\alias{random_encouragement}
\title{Random praise and encouragement}
\usage{
random_praise()

random_encouragement()
}
\value{
Character string with a random saying
}
\description{
Random praises and encouragements sayings to compliment your question and quiz experience.
}
