% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_exercise.R
\name{run_exercise}
\alias{run_exercise}
\title{Run an exercise and return it's results}
\usage{
run_exercise(input, output)
}
\arguments{
\item{input}{Connection or path to file containing input JSON}

\item{output}{Connection or path to file to write output JSON}
}
\description{
Run a JSON serialiazed exercise and return the results as JSON.
}

