% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{safe_env}
\alias{safe_env}
\title{Safe R CMD environment}
\usage{
safe_env()
}
\description{
By default, \code{callr::\link[callr]{rcmd_safe_env}} suppresses the ability
to open a browser window.  This is the default execution environment within
\code{callr::\link[callr]{r}}.  However, opening a browser is expected
behavior within the learnr package and should not be suppressed.
}
