% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-dependencies.R
\name{tutorial_html_dependency}
\alias{tutorial_html_dependency}
\title{Tutorial HTML dependency}
\usage{
tutorial_html_dependency()
}
\description{
Tutorial HTML dependency
}
\details{
HTML dependency for core tutorial JS and CSS. This should be included as a
dependency for custom tutorial formats that wish to ensure that that
tutorial.js and tutorial.css are loaded prior their own scripts and stylesheets.
}
