% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tutorial_options}
\alias{tutorial_options}
\title{Set tutorial options}
\usage{
tutorial_options(
  exercise.cap = "Code",
  exercise.eval = FALSE,
  exercise.timelimit = 30,
  exercise.lines = NULL,
  exercise.checker = NULL,
  exercise.completion = TRUE,
  exercise.diagnostics = TRUE,
  exercise.startover = TRUE
)
}
\arguments{
\item{exercise.cap}{Caption for exercise chunk (defaults to \code{"Code")}.}

\item{exercise.eval}{Whether to pre-evaluate the exercise so the reader can
see some default output (defaults to \code{FALSE}).}

\item{exercise.timelimit}{Number of seconds to limit execution time to
(defaults to \code{30}).}

\item{exercise.lines}{Lines of code for exercise editor (defaults to the
number of lines in the code chunk).}

\item{exercise.checker}{Function used to check exercise answers.}

\item{exercise.completion}{Use code completion in exercise editors.}

\item{exercise.diagnostics}{Show diagnostics in exercise editors.}

\item{exercise.startover}{Show "Start Over" button on exercise.}
}
\description{
Set various tutorial options that control the display and evaluation of
exercises.
}
