% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_package_dependencies.R
\name{tutorial_package_dependencies}
\alias{tutorial_package_dependencies}
\title{List Tutorial Dependencies}
\usage{
tutorial_package_dependencies(name = NULL, package = NULL)
}
\arguments{
\item{name}{The tutorial name. If \code{name} is \code{NULL}, then all
tutorials within \code{package} will be searched.}

\item{package}{The \R package providing the tutorial. If \code{package} is
\code{NULL}, then all tutorials will be searched.}
}
\value{
A character vector of package names that are required for execution.
}
\description{
List the \R packages required to run a particular tutorial.
}
\examples{
tutorial_package_dependencies(package = "learnr")
}
